requirements.txt
django >4.2,<4.3
python-decouple==3.8
Pillow


djangorestframework
markdown       
django-filter  